requirements.txt
docopt==0.6.2
ffmpeg-python==0.2.0
future==0.18.2
MIDIUtil==1.2.1
mutagen==1.45.1
numpy==1.21.5
pyFluidSynth==1.3.0
